+++ tests/test_examples.py
@@ -71,6 +71,22 @@ class TestExamples(unittest.TestCase):
             until=8640,
             with_plots=False,
             log_level=logging.DEBUG,
+            use_direct_callback_databroker=False
+        )
+
+    def test_room_mas_direct_callback(self):
+        """Test the room_mas example"""
+        import sys
+
+        if "linux" in sys.platform:
+            self.skipTest("FMU test not yet supported")
+        self._run_example_with_return(
+            file="multi-agent-systems//room_mas//room_mas.py",
+            func_name="run_example",
+            until=8640,
+            with_plots=False,
+            log_level=logging.DEBUG,
+            use_direct_callback_databroker=True
         )
 
     def test_pingpong(self):